+++ examples/basic_examples/block_cpu.c
@@ -19,13 +19,13 @@
 void cpu_codelet(void *descr[], void *_args)
 {
 	float *block = (float *)STARPU_BLOCK_GET_PTR(descr[0]);
-	int nx = (int)STARPU_BLOCK_GET_NX(descr[0]);
-	int ny = (int)STARPU_BLOCK_GET_NY(descr[0]);
-	int nz = (int)STARPU_BLOCK_GET_NZ(descr[0]);
-	unsigned ldy = STARPU_BLOCK_GET_LDY(descr[0]);
-	unsigned ldz = STARPU_BLOCK_GET_LDZ(descr[0]);
+	size_t nx = STARPU_BLOCK_GET_NX(descr[0]);
+	size_t ny = STARPU_BLOCK_GET_NY(descr[0]);
+	size_t nz = STARPU_BLOCK_GET_NZ(descr[0]);
+	size_t ldy = STARPU_BLOCK_GET_LDY(descr[0]);
+	size_t ldz = STARPU_BLOCK_GET_LDZ(descr[0]);
 	float *multiplier = (float *)_args;
-	int i, j, k;
+	size_t i, j, k;
 
 	for(k=0; k<nz ; k++)
 	{
