+++ examples/basic_examples/vector_scal_cpu_template.h
@@ -30,7 +30,7 @@
 #define VECTOR_SCAL_CPU_FUNC(func_name)                                        \
 void func_name(void *buffers[], void *cl_arg)                                  \
 {                                                                              \
-	unsigned i;                                                            \
+	size_t i; 							       \
 	float *factor = (float *) cl_arg;                                      \
 									       \
 	/*                                                                     \
@@ -47,7 +47,7 @@ void func_name(void *buffers[], void *cl_arg)                                  \
 	struct starpu_vector_interface *vector = (struct starpu_vector_interface *) buffers[0]; \
 									       \
 	/* length of the vector */                                             \
-	unsigned n = STARPU_VECTOR_GET_NX(vector);                             \
+	size_t n = STARPU_VECTOR_GET_NX(vector);			\
 									       \
 	/* get a pointer to the local copy of the vector : note that we have to\
 	 * cast it in (float *) since a vector could contain any type of       \
@@ -64,22 +64,22 @@ void func_name(void *buffers[], void *cl_arg)                                  \
 void func_name(void *buffers[], void *cl_arg)                                  \
 {                                                                              \
 	float *vector = (float *) STARPU_VECTOR_GET_PTR(buffers[0]);           \
-	unsigned int n = STARPU_VECTOR_GET_NX(buffers[0]);                     \
-	unsigned int n_iterations = n/4;                                       \
+	size_t n = STARPU_VECTOR_GET_NX(buffers[0]);                           \
+	size_t n_iterations = n/4;                                             \
 									       \
 	__m128 *VECTOR = (__m128*) vector;                                     \
 	__m128 FACTOR STARPU_ATTRIBUTE_ALIGNED(16);                            \
 	float factor = *(float *) cl_arg;                                      \
 	FACTOR = _mm_set1_ps(factor);                                          \
 									       \
-	unsigned int i;	                                                       \
+	size_t i;	                                                       \
 	for (i = 0; i < n_iterations; i++)                                     \
 		VECTOR[i] = _mm_mul_ps(FACTOR, VECTOR[i]);                     \
 									       \
 	unsigned int remainder = n%4;                                          \
 	if (remainder != 0)                                                    \
 	{                                                                      \
-		unsigned int start = 4 * n_iterations;                         \
+		size_t start = 4 * n_iterations;                               \
 		for (i = start; i < start+remainder; ++i)                      \
 		{                                                              \
 			vector[i] = factor * vector[i];                        \
