+++ examples/basic_examples/vector_scal_opencl_kernel.cl
@@ -14,7 +14,7 @@
  * See the GNU Lesser General Public License in COPYING.LGPL for more details.
  */
 
-__kernel void vector_mult_opencl(unsigned int nx, __global float* val, float factor)
+__kernel void vector_mult_opencl(ulong nx, __global float* val, float factor)
 {
         const int i = get_global_id(0);
         if (i < nx)
